put ModelFileExtension in its own file
