+++ Sources/CoreDataModel.swift
@@ -20,25 +20,6 @@ import CoreData
 import Foundation
 
 /**
- Describes a Core Data model file exention type based on the
- [Model File Format and Versions](https://developer.apple.com/library/ios/documentation/Cocoa/Conceptual/CoreDataVersioning/Articles/vmModelFormat.html)
- documentation.
- */
-public enum ModelFileExtension: String {
-    /// The extension for a model bundle, or a `.xcdatamodeld` file package.
-    case bundle = "momd"
-
-    /// The extension for a versioned model file, or a `.xcdatamodel` file.
-    case versionedFile = "mom"
-
-    /// The extension for a mapping model file, or a `.xcmappingmodel` file.
-    case mapping = "cdm"
-
-    /// The extension for a sqlite store.
-    case sqlite = "sqlite"
-}
-
-/**
  An instance of `CoreDataModel` represents a Core Data model — a `.xcdatamodeld` file package.
  It provides the model and store URLs as well as methods for interacting with the store.
  */
